Renamed to interfaces.ts and fixed typo
